refactor(ruby/x402): mirror Rust spine layout
Restructure ruby/lib/x402/ to mirror the Rust spine at
rust/crates/x402/src/ instead of the previous interop-flavored
single namespace.

  lib/x402.rb                                  -> lib.rs
  lib/x402/constants.rb                        -> constants.rs
  lib/x402/error.rb                            -> error.rs
  lib/x402/protocol/schemes/exact/types.rb     -> protocol/schemes/exact/types.rs
  lib/x402/protocol/schemes/exact/verify.rb    -> protocol/schemes/exact/verify.rs
  lib/x402/server/exact.rb                     -> server/exact.rs
  bin/x402-interop-server                      -> bin/interop_server.rs

X402::Server::Exact is the production server entry point; the
former X402::Interop::Server::State becomes X402::Server::Exact::Config
(State alias retained for back-compat). The 11-rule verifier moves
into X402::Protocol::Schemes::Exact::Verifier with each rule citing
the spine verify.rs line range. The interop bin is a thin TCP
adapter; all harness env reads (X402_INTEROP_*) live in the bin,
not in the library.
