+++ ruby/test/x402_server_exact_test.rb
@@ -3,10 +3,9 @@
 require "base64"
 require "json"
 require_relative "test_helper"
-require "x402/interop/exact"
-require "x402/interop/server"
+require "x402"
 
-class InteropServerTest < Minitest::Test
+class X402ServerExactTest < Minitest::Test
   NETWORK = "solana:EtWTRABZaYq6iMfeYKouRu166VU2xqa1"
   ASSET = "4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU"
   EXTRA_ASSET = "ExtraMint11111111111111111111111111111"
@@ -15,27 +14,27 @@ class InteropServerTest < Minitest::Test
   BLOCKHASH = "11111111111111111111111111111111"
 
   def test_normalizes_price_to_six_decimals
-    assert_equal "1000", X402::Interop::Server.normalize_amount("$0.001")
-    assert_equal "1000", X402::Interop::Server.normalize_amount("0.001 USDC")
-    assert_equal "1250000", X402::Interop::Server.normalize_amount("1.25")
+    assert_equal "1000", X402::Server::Exact.normalize_amount("$0.001")
+    assert_equal "1000", X402::Server::Exact.normalize_amount("0.001 USDC")
+    assert_equal "1250000", X402::Server::Exact.normalize_amount("1.25")
   end
 
   def test_exact_challenge_uses_runtime_state
     state = build_state(price: "$0.125")
-    requirement = X402::Interop::Server.exact_requirement(state)
+    requirement = X402::Server::Exact.exact_requirement(state)
 
     assert_equal "exact", requirement.fetch("scheme")
     assert_equal NETWORK, requirement.fetch("network")
     assert_equal ASSET, requirement.fetch("asset")
     assert_equal "125000", requirement.fetch("amount")
     assert_equal PAY_TO, requirement.fetch("payTo")
-    assert_equal X402::Interop::Exact.base58_encode(state.fee_payer.raw_public_key),
+    assert_equal X402::Protocol::Schemes::Exact.base58_encode(state.fee_payer.raw_public_key),
       requirement.fetch("extra").fetch("feePayer")
   end
 
   def test_exact_challenge_includes_extra_offered_mints
     state = build_state(extra_offered_mints: " #{PYUSD_DEVNET_MINT}, #{EXTRA_ASSET} ")
-    accepts = X402::Interop::Server.exact_challenge(state).fetch("accepts")
+    accepts = X402::Server::Exact.exact_challenge(state).fetch("accepts")
     base, pyusd, extra = accepts
 
     assert_equal [ASSET, PYUSD_DEVNET_MINT, EXTRA_ASSET], accepts.map { |requirement| requirement.fetch("asset") }
@@ -48,20 +47,20 @@ class InteropServerTest < Minitest::Test
       assert_equal base.fetch("extra").fetch("decimals"), requirement.fetch("extra").fetch("decimals")
     end
 
-    assert_equal X402::Interop::Exact::TOKEN_2022_PROGRAM, pyusd.fetch("extra").fetch("tokenProgram")
-    assert_equal X402::Interop::Server::DEFAULT_TOKEN_PROGRAM, extra.fetch("extra").fetch("tokenProgram")
+    assert_equal X402::Protocol::Schemes::Exact::TOKEN_2022_PROGRAM, pyusd.fetch("extra").fetch("tokenProgram")
+    assert_equal X402::Server::Exact::DEFAULT_TOKEN_PROGRAM, extra.fetch("extra").fetch("tokenProgram")
   end
 
   def test_payment_requirement_matches_binds_settlement_fields
     state = build_state
-    requirement = X402::Interop::Server.exact_requirement(state)
+    requirement = X402::Server::Exact.exact_requirement(state)
 
-    assert X402::Interop::Server.payment_requirement_matches?(requirement, requirement)
+    assert X402::Server::Exact.payment_requirement_matches?(requirement, requirement)
 
     mutated = Marshal.load(Marshal.dump(requirement))
     mutated.fetch("extra")["feePayer"] = "11111111111111111111111111111114"
 
-    refute X402::Interop::Server.payment_requirement_matches?(mutated, requirement)
+    refute X402::Server::Exact.payment_requirement_matches?(mutated, requirement)
   end
 
   def test_settlement_signs_fee_payer_before_sending
@@ -72,7 +71,7 @@ class InteropServerTest < Minitest::Test
     })
     payment_header = build_payment_header(state)
 
-    settlement = X402::Interop::Server.settle_exact_payment(state, payment_header)
+    settlement = X402::Server::Exact.settle_exact_payment(state, payment_header)
     signed_transaction = sent.fetch(0)
 
     assert_equal "ruby-settlement-signature", settlement
@@ -87,7 +86,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "No matching payment requirements: accepted payment requirement does not match server challenge", error.message
@@ -100,7 +99,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "No matching payment requirements: accepted payment requirement does not match server challenge", error.message
@@ -113,7 +112,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "No matching payment requirements: accepted payment requirement does not match server challenge", error.message
@@ -123,7 +122,7 @@ class InteropServerTest < Minitest::Test
     state = build_state
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, "not base64")
+      X402::Server::Exact.settle_exact_payment(state, "not base64")
     end
 
     assert_equal "invalid payment signature encoding", error.message
@@ -134,7 +133,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64("not-json")
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid payment signature JSON", error.message
@@ -145,7 +144,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64(JSON.generate(["not", "object"]))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "payment signature must be a JSON object", error.message
@@ -155,13 +154,13 @@ class InteropServerTest < Minitest::Test
     state = build_state
     envelope = {
       "x402Version" => 2,
-      "accepted" => X402::Interop::Server.exact_requirement(state),
+      "accepted" => X402::Server::Exact.exact_requirement(state),
       "payload" => "not-object"
     }
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "payment payload is missing transaction", error.message
@@ -171,13 +170,13 @@ class InteropServerTest < Minitest::Test
     state = build_state
     envelope = {
       "x402Version" => 2,
-      "accepted" => X402::Interop::Server.exact_requirement(state),
+      "accepted" => X402::Server::Exact.exact_requirement(state),
       "payload" => {}
     }
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "payment payload is missing transaction", error.message
@@ -190,7 +189,7 @@ class InteropServerTest < Minitest::Test
     payment_header = Base64.strict_encode64(JSON.generate(envelope))
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "payment payload transaction is not valid base64", error.message
@@ -207,7 +206,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_amount_mismatch", error.message
@@ -225,7 +224,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_transferring_funds", error.message
@@ -243,7 +242,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_transferring_funds", error.message
@@ -261,7 +260,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_in_instruction_accounts", error.message
@@ -286,7 +285,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_in_instruction_accounts", error.message
@@ -308,7 +307,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_in_instruction_accounts", error.message
@@ -330,7 +329,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_transaction_fee_payer_in_instruction_accounts", error.message
@@ -344,7 +343,7 @@ class InteropServerTest < Minitest::Test
     state = build_state(sender: ->(_state, _transaction) { "unit-settlement" })
 
     assert_equal "unit-settlement",
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
   end
 
   def test_settlement_rejects_lighthouse_as_sixth_instruction
@@ -354,12 +353,12 @@ class InteropServerTest < Minitest::Test
       "unit-settlement"
     })
     payment_header = mutate_payment_transaction(build_payment_header(state)) do |transaction|
-      append_optional_instruction(transaction, X402::Interop::Exact::LIGHTHOUSE_PROGRAM)
-      append_optional_instruction(transaction, X402::Interop::Exact::LIGHTHOUSE_PROGRAM)
+      append_optional_instruction(transaction, X402::Protocol::Schemes::Exact::LIGHTHOUSE_PROGRAM)
+      append_optional_instruction(transaction, X402::Protocol::Schemes::Exact::LIGHTHOUSE_PROGRAM)
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_unknown_sixth_instruction", error.message
@@ -375,9 +374,9 @@ class InteropServerTest < Minitest::Test
     state = build_state(sender: ->(_state, _transaction) { "shared-signature" })
     payment_header = build_payment_header(state)
 
-    assert_equal "shared-signature", X402::Interop::Server.settle_exact_payment(state, payment_header)
+    assert_equal "shared-signature", X402::Server::Exact.settle_exact_payment(state, payment_header)
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "signature_consumed", error.message
@@ -385,7 +384,7 @@ class InteropServerTest < Minitest::Test
 
   def test_settlement_orders_broadcast_then_confirm_then_put_if_absent
     order = []
-    cache = X402::Interop::Server::SettlementCache.new
+    cache = X402::Server::Exact::SettlementCache.new
     tracking_cache = Class.new do
       def initialize(inner, order)
         @inner = inner
@@ -414,7 +413,7 @@ class InteropServerTest < Minitest::Test
     )
 
     assert_equal "sig-ordering",
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
 
     assert_equal [
       [:broadcast],
@@ -424,7 +423,7 @@ class InteropServerTest < Minitest::Test
   end
 
   def test_settlement_does_not_record_signature_when_broadcast_fails_before_confirm
-    cache = X402::Interop::Server::SettlementCache.new
+    cache = X402::Server::Exact::SettlementCache.new
     state = build_state(
       sender: ->(_state, _transaction) { raise "sendTransaction RPC error: blockhash not found" },
       signature_confirmer: ->(_state, _signature) { raise "confirm must not run when broadcast failed" },
@@ -433,7 +432,7 @@ class InteropServerTest < Minitest::Test
     payment_header = build_payment_header(state)
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
     assert_match(/blockhash not found/, error.message)
 
@@ -448,12 +447,12 @@ class InteropServerTest < Minitest::Test
       signature_confirmer: ->(_state, signature) { signature },
       settlement_cache: cache
     )
-    assert_equal "retry-sig", X402::Interop::Server.settle_exact_payment(state, payment_header)
+    assert_equal "retry-sig", X402::Server::Exact.settle_exact_payment(state, payment_header)
     assert retried
   end
 
   def test_settlement_does_not_record_signature_when_confirmation_fails
-    cache = X402::Interop::Server::SettlementCache.new
+    cache = X402::Server::Exact::SettlementCache.new
     state = build_state(
       sender: ->(_state, _transaction) { "unconfirmed-sig" },
       signature_confirmer: ->(_state, _signature) { raise "timed out awaiting confirmation for unconfirmed-sig" },
@@ -461,7 +460,7 @@ class InteropServerTest < Minitest::Test
     )
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
     end
     assert_match(/timed out awaiting confirmation/, error.message)
 
@@ -474,7 +473,7 @@ class InteropServerTest < Minitest::Test
 
   def test_settlement_consumed_key_namespace_is_scheme_scoped
     assert_equal "x402-svm-exact:consumed:abc123",
-      X402::Interop::Server.signature_consumed_key("abc123")
+      X402::Server::Exact.signature_consumed_key("abc123")
   end
 
   def test_settlement_rejects_missing_source_token_account_before_sending
@@ -492,7 +491,7 @@ class InteropServerTest < Minitest::Test
     )
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
     end
 
     assert_equal "source token account does not exist", error.message
@@ -515,7 +514,7 @@ class InteropServerTest < Minitest::Test
     )
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
     end
 
     assert_equal "destination token account does not exist", error.message
@@ -535,7 +534,7 @@ class InteropServerTest < Minitest::Test
       append_valid_destination_ata_create_instruction(transaction, state)
     end
 
-    assert_equal "unit-settlement", X402::Interop::Server.settle_exact_payment(state, payment_header)
+    assert_equal "unit-settlement", X402::Server::Exact.settle_exact_payment(state, payment_header)
     assert_equal 1, checked.length
   end
 
@@ -561,7 +560,7 @@ class InteropServerTest < Minitest::Test
     end
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header)
+      X402::Server::Exact.settle_exact_payment(state, payment_header)
     end
 
     assert_equal "invalid_exact_svm_payload_signature", error.message
@@ -579,7 +578,7 @@ class InteropServerTest < Minitest::Test
     state = build_state(sender: ->(_state, _transaction) { "unit-settlement" })
 
     assert_equal "unit-settlement",
-      X402::Interop::Server.settle_exact_payment(state, build_payment_header(state))
+      X402::Server::Exact.settle_exact_payment(state, build_payment_header(state))
   end
 
   def test_server_rejects_payment_for_different_resource
@@ -587,7 +586,7 @@ class InteropServerTest < Minitest::Test
     payment_header = build_payment_header(state, resource: "/resource/a")
 
     error = assert_raises(RuntimeError) do
-      X402::Interop::Server.settle_exact_payment(state, payment_header, resource: "/resource/b")
+      X402::Server::Exact.settle_exact_payment(state, payment_header, resource: "/resource/b")
     end
 
     assert_equal "invalid_exact_svm_payload_resource_mismatch", error.message
@@ -598,11 +597,11 @@ class InteropServerTest < Minitest::Test
     payment_header = build_payment_header(state, resource: "/resource/a")
 
     assert_equal "unit-settlement",
-      X402::Interop::Server.settle_exact_payment(state, payment_header, resource: "/resource/a")
+      X402::Server::Exact.settle_exact_payment(state, payment_header, resource: "/resource/a")
   end
 
   def test_settlement_cache_evicts_entries_after_ttl
-    cache = X402::Interop::Server::SettlementCache.new(ttl_seconds: 120)
+    cache = X402::Server::Exact::SettlementCache.new(ttl_seconds: 120)
     now = Time.at(1_000)
 
     refute cache.duplicate?("tx-a", now: now)
@@ -611,7 +610,7 @@ class InteropServerTest < Minitest::Test
   end
 
   def test_payment_errors_are_normalized
-    body = X402::Interop::Server.payment_error_body(RuntimeError.new("sendTransaction RPC error: failed"))
+    body = X402::Server::Exact.payment_error_body(RuntimeError.new("sendTransaction RPC error: failed"))
 
     assert_equal(
       {
@@ -625,7 +624,7 @@ class InteropServerTest < Minitest::Test
 
   def test_protected_route_normalizes_invalid_payment_error_body
     state = build_state
-    status, headers, body = X402::Interop::Server.response_for(
+    status, headers, body = X402::Server::Exact.response_for(
       "/protected",
       {"PAYMENT-SIGNATURE" => "not base64"},
       state
@@ -661,7 +660,7 @@ class InteropServerTest < Minitest::Test
     singleton.define_method(:start, start)
     begin
       error = assert_raises(RuntimeError) do
-        X402::Interop::Server.send_transaction(state, "signed-transaction")
+        X402::Server::Exact.send_transaction(state, "signed-transaction")
       end
 
       assert_equal "sendTransaction RPC error: Transaction simulation failed", error.message
@@ -674,7 +673,7 @@ class InteropServerTest < Minitest::Test
     state = build_state
 
     with_net_http_response(JSON.generate("result" => "rpc-signature")) do
-      assert_equal "rpc-signature", X402::Interop::Server.send_transaction(state, "signed-transaction")
+      assert_equal "rpc-signature", X402::Server::Exact.send_transaction(state, "signed-transaction")
     end
   end
 
@@ -683,7 +682,7 @@ class InteropServerTest < Minitest::Test
 
     with_net_http_response(JSON.generate("result" => "")) do
       error = assert_raises(RuntimeError) do
-        X402::Interop::Server.send_transaction(state, "signed-transaction")
+        X402::Server::Exact.send_transaction(state, "signed-transaction")
       end
 
       assert_equal "sendTransaction returned empty signature", error.message
@@ -694,7 +693,7 @@ class InteropServerTest < Minitest::Test
     state = build_state
 
     with_net_http_response(JSON.generate("result" => {"value" => {"owner" => "token"}})) do
-      assert X402::Interop::Server.account_exists?(state, PAY_TO)
+      assert X402::Server::Exact.account_exists?(state, PAY_TO)
     end
   end
 
@@ -702,7 +701,7 @@ class InteropServerTest < Minitest::Test
     state = build_state
 
     with_net_http_response(JSON.generate("result" => {"value" => nil})) do
-      refute X402::Interop::Server.account_exists?(state, PAY_TO)
+      refute X402::Server::Exact.account_exists?(state, PAY_TO)
     end
   end
 
@@ -711,7 +710,7 @@ class InteropServerTest < Minitest::Test
 
     with_net_http_response(JSON.generate("error" => "plain rpc failure")) do
       error = assert_raises(RuntimeError) do
-        X402::Interop::Server.account_exists?(state, PAY_TO)
+        X402::Server::Exact.account_exists?(state, PAY_TO)
       end
 
       assert_equal "getAccountInfo RPC error: plain rpc failure", error.message
@@ -723,7 +722,7 @@ class InteropServerTest < Minitest::Test
 
     with_net_http_response("service unavailable", code: "503", success: false) do
       error = assert_raises(RuntimeError) do
-        X402::Interop::Server.account_exists?(state, PAY_TO)
+        X402::Server::Exact.account_exists?(state, PAY_TO)
       end
 
       assert_equal "getAccountInfo HTTP 503", error.message
@@ -733,19 +732,19 @@ class InteropServerTest < Minitest::Test
   def test_static_routes_return_expected_responses
     state = build_state
 
-    status, = X402::Interop::Server.response_for("/health", {}, state)
+    status, = X402::Server::Exact.response_for("/health", {}, state)
     assert_equal 200, status
 
-    status, _headers, body = X402::Interop::Server.response_for("/capabilities", {}, state)
+    status, _headers, body = X402::Server::Exact.response_for("/capabilities", {}, state)
     assert_equal 200, status
     assert_equal "ruby", body.fetch(:implementation)
 
-    status, headers, body = X402::Interop::Server.response_for("/exact", {}, state)
+    status, headers, body = X402::Server::Exact.response_for("/exact", {}, state)
     assert_equal 402, status
     assert headers.key?("PAYMENT-REQUIRED")
     assert_equal({error: "payment_required"}, body)
 
-    status, headers, body = X402::Interop::Server.response_for("/missing", {}, state)
+    status, headers, body = X402::Server::Exact.response_for("/missing", {}, state)
     assert_equal 404, status
     assert_empty headers
     assert_equal({error: "not_found"}, body)
@@ -753,7 +752,7 @@ class InteropServerTest < Minitest::Test
 
   def test_protected_route_returns_settlement_success
     state = build_state(sender: ->(_state, _transaction) { "settlement-signature" })
-    status, headers, body = X402::Interop::Server.response_for(
+    status, headers, body = X402::Server::Exact.response_for(
       "/protected",
       {"payment-signature" => build_payment_header(state, resource: "/protected")},
       state
@@ -792,14 +791,14 @@ class InteropServerTest < Minitest::Test
       "X402_INTEROP_FACILITATOR_SECRET_KEY" => JSON.generate(secret(65)),
       "X402_INTEROP_PRICE" => "$0.001"
     }
-    server_b = X402::Interop::Server::State.new(
+    server_b = X402::Server::Exact::Config.new(
       env: server_b_env,
       transaction_sender: ->(_state, _transaction) { "settlement-signature" },
       account_checker: ->(_state, _account) { true }
     )
     payment_header = build_payment_header(server_a, resource: "/protected")
 
-    status, _headers, body = X402::Interop::Server.response_for(
+    status, _headers, body = X402::Server::Exact.response_for(
       "/protected",
       {"PAYMENT-SIGNATURE" => payment_header},
       server_b
@@ -817,7 +816,7 @@ class InteropServerTest < Minitest::Test
 
   def test_protected_route_returns_payment_required_without_signature
     state = build_state
-    status, headers, body = X402::Interop::Server.response_for("/protected", {}, state)
+    status, headers, body = X402::Server::Exact.response_for("/protected", {}, state)
 
     assert_equal 402, status
     assert_equal({error: "payment_required"}, body)
@@ -836,19 +835,19 @@ class InteropServerTest < Minitest::Test
     )
 
     # Default route no longer routes here.
-    status, _headers, body = X402::Interop::Server.response_for("/protected", {}, state)
+    status, _headers, body = X402::Server::Exact.response_for("/protected", {}, state)
     assert_equal 404, status
     assert_equal({error: "not_found"}, body)
 
     # Challenge advertises the overridden resource URI.
-    status, headers, _body = X402::Interop::Server.response_for("/protected/expensive", {}, state)
+    status, headers, _body = X402::Server::Exact.response_for("/protected/expensive", {}, state)
     assert_equal 402, status
     challenge = JSON.parse(Base64.decode64(headers.fetch("PAYMENT-REQUIRED")))
     assert_equal "/protected/expensive", challenge.fetch("resource").fetch("uri")
 
     # Settlement emits the overridden header name and not the default.
     payment_header = build_payment_header(state, resource: "/protected/expensive")
-    status, headers, body = X402::Interop::Server.response_for(
+    status, headers, body = X402::Server::Exact.response_for(
       "/protected/expensive",
       {"PAYMENT-SIGNATURE" => payment_header},
       state
@@ -872,7 +871,7 @@ class InteropServerTest < Minitest::Test
       "X402_INTEROP_RESOURCE_PATH" => resource_path,
       "X402_INTEROP_SETTLEMENT_HEADER" => settlement_header
     }
-    X402::Interop::Server::State.new(
+    X402::Server::Exact::Config.new(
       env: env,
       transaction_sender: sender,
       account_checker: ->(_state, _account) { true },
@@ -898,7 +897,7 @@ class InteropServerTest < Minitest::Test
     }
     env["X402_INTEROP_EXTRA_OFFERED_MINTS"] = extra_offered_mints unless extra_offered_mints.nil?
 
-    X402::Interop::Server::State.new(
+    X402::Server::Exact::Config.new(
       env: env,
       transaction_sender: sender,
       account_checker: account_checker,
@@ -908,8 +907,8 @@ class InteropServerTest < Minitest::Test
   end
 
   def build_payment_header(state, resource: nil)
-    X402::Interop::Exact.build_exact_payment_signature(
-      requirement: X402::Interop::Server.exact_requirement(state, resource: resource),
+    X402::Protocol::Schemes::Exact.build_exact_payment_signature(
+      requirement: X402::Server::Exact.exact_requirement(state, resource: resource),
       client_secret_key: JSON.generate(secret(1)),
       recent_blockhash: BLOCKHASH,
       resource: {"type" => "http", "uri" => resource || "/protected"}
@@ -946,10 +945,10 @@ class InteropServerTest < Minitest::Test
 
   def resign_client_signature(transaction)
     bytes = transaction.b
-    signature_count, signatures_offset = X402::Interop::Exact.read_short_vec(bytes, 0)
+    signature_count, signatures_offset = X402::Protocol::Schemes::Exact.read_short_vec(bytes, 0)
     message_offset = signatures_offset + (signature_count * 64)
     message = bytes.byteslice(message_offset, bytes.bytesize - message_offset)
-    private_key = X402::Interop::Exact.private_key_from_json(JSON.generate(secret(1)))
+    private_key = X402::Protocol::Schemes::Exact.private_key_from_json(JSON.generate(secret(1)))
     # Client signer is at index 1 (fee_payer is 0).
     signature = private_key.sign(nil, message)
     bytes[signatures_offset + 64, 64] = signature
@@ -989,9 +988,9 @@ class InteropServerTest < Minitest::Test
     account_keys_offset = account_count_offset + 1
     blockhash_offset = account_keys_offset + (account_count * 32)
 
-    unless transaction.byteslice(account_keys_offset, account_count * 32).include?(X402::Interop::Exact.base58_decode(program))
+    unless transaction.byteslice(account_keys_offset, account_count * 32).include?(X402::Protocol::Schemes::Exact.base58_decode(program))
       transaction.setbyte(account_count_offset, account_count + 1)
-      transaction.insert(blockhash_offset, X402::Interop::Exact.base58_decode(program))
+      transaction.insert(blockhash_offset, X402::Protocol::Schemes::Exact.base58_decode(program))
       account_count += 1
     end
 
@@ -1009,9 +1008,9 @@ class InteropServerTest < Minitest::Test
     account_keys_offset = account_count_offset + 1
     blockhash_offset = account_keys_offset + (account_count * 32)
     extra_keys = [
-      X402::Interop::Exact.base58_decode(state.pay_to),
-      X402::Interop::Exact.base58_decode(X402::Interop::Exact::SYSTEM_PROGRAM),
-      X402::Interop::Exact.base58_decode(X402::Interop::Exact::ASSOCIATED_TOKEN_PROGRAM)
+      X402::Protocol::Schemes::Exact.base58_decode(state.pay_to),
+      X402::Protocol::Schemes::Exact.base58_decode(X402::Protocol::Schemes::Exact::SYSTEM_PROGRAM),
+      X402::Protocol::Schemes::Exact.base58_decode(X402::Protocol::Schemes::Exact::ASSOCIATED_TOKEN_PROGRAM)
     ]
 
     transaction.setbyte(account_count_offset, account_count + extra_keys.length)
@@ -1077,7 +1076,7 @@ class InteropServerTest < Minitest::Test
 
     # Add SystemProgram as a new static account key.
     transaction.setbyte(account_count_offset, account_count + 1)
-    transaction.insert(blockhash_offset, X402::Interop::Exact.base58_decode(X402::Interop::Exact::SYSTEM_PROGRAM))
+    transaction.insert(blockhash_offset, X402::Protocol::Schemes::Exact.base58_decode(X402::Protocol::Schemes::Exact::SYSTEM_PROGRAM))
     system_program_index = account_count
 
     new_account_count = account_count + 1
